Make batch command top level
