+++ tools/cli/workflow.go
@@ -28,8 +28,6 @@ import (
 	"strings"
 
 	"github.com/urfave/cli"
-
-	"github.com/temporalio/temporal/service/worker/batcher"
 )
 
 func newWorkflowCommands() []cli.Command {
@@ -323,101 +321,6 @@ func newWorkflowCommands() []cli.Command {
 			},
 			Action: func(c *cli.Context) {
 				ResetInBatch(c)
-			},
-		},
-		{
-			Name:        "batch",
-			Usage:       "batch operation on a list of workflows from query.",
-			Subcommands: newBatchCommands(),
-		},
-	}
-}
-
-func newBatchCommands() []cli.Command {
-	return []cli.Command{
-		{
-			Name:    "describe",
-			Aliases: []string{"desc"},
-			Usage:   "Describe a batch operation job",
-			Flags: []cli.Flag{
-				cli.StringFlag{
-					Name:  FlagJobIDWithAlias,
-					Usage: "Batch Job Id",
-				},
-			},
-			Action: func(c *cli.Context) {
-				DescribeBatchJob(c)
-			},
-		},
-		{
-			Name:  "terminate",
-			Usage: "terminate a batch operation job",
-			Flags: []cli.Flag{
-				cli.StringFlag{
-					Name:  FlagJobIDWithAlias,
-					Usage: "Batch Job Id",
-				},
-				cli.StringFlag{
-					Name:  FlagReasonWithAlias,
-					Usage: "Reason to stop this batch job",
-				},
-			},
-			Action: func(c *cli.Context) {
-				TerminateBatchJob(c)
-			},
-		},
-		{
-			Name:    "list",
-			Aliases: []string{"l"},
-			Usage:   "Describe a batch operation job",
-			Flags: []cli.Flag{
-				cli.IntFlag{
-					Name:  FlagPageSizeWithAlias,
-					Value: 30,
-					Usage: "Result page size",
-				},
-			},
-			Action: func(c *cli.Context) {
-				ListBatchJobs(c)
-			},
-		},
-		{
-			Name:  "start",
-			Usage: "Start a batch operation job",
-			Flags: []cli.Flag{
-				cli.StringFlag{
-					Name:  FlagListQueryWithAlias,
-					Usage: "Query to get workflows for being executed this batch operation",
-				},
-				cli.StringFlag{
-					Name:  FlagReasonWithAlias,
-					Usage: "Reason to run this batch job",
-				},
-				cli.StringFlag{
-					Name:  FlagBatchTypeWithAlias,
-					Usage: "Types supported: " + strings.Join(batcher.AllBatchTypes, ","),
-				},
-				//below are optional
-				cli.StringFlag{
-					Name:  FlagSignalNameWithAlias,
-					Usage: "Required for batch signal",
-				},
-				cli.StringFlag{
-					Name:  FlagInputWithAlias,
-					Usage: "Optional input of signal",
-				},
-				cli.IntFlag{
-					Name:  FlagRPS,
-					Value: batcher.DefaultRPS,
-					Usage: "RPS of processing",
-				},
-				cli.BoolFlag{
-					Name:  FlagYes,
-					Usage: "Optional flag to disable confirmation prompt",
-				},
-			},
-			Action: func(c *cli.Context) {
-				StartBatchJob(c)
 			},
 		},
 	}
